requirements.txt
streamlit
pandas
fpdf
psycopg2-binary
openpyxl